process_tracker_python-62 Refactor data_store
Moved data_store into utilities.  Cleaned up config in data_store.

Closes #62
