+++ tests/utilities/test_utilities.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta
 import unittest
 
-from process_tracker.data_store import DataStore
+from process_tracker.utilities.data_store import DataStore
 from process_tracker.utilities import utilities
